refactor(WS2-AC4): relocate v2RoutingTable literal → RoutingTable.swift
Pure move of the 250-line routing-table static let into a new RoutingTable.swift as extension ChannelRouter (same name/type/visibility). ChannelRouter.v2RoutingTable access unchanged; golden route map diff = 0.
